Use same leader for metrics & purge
