+++ go/database/log/queries/queries.sql
@@ -35,4 +35,4 @@ INSERT INTO collection (id, record_enumeration_offset_position, record_compactio
 DELETE FROM record_log r using collection c where r.collection_id = c.id and r.offset <= c.record_compaction_offset_position;
 
 -- name: GetTotalUncompactedRecordsCount :one
-SELECT SUM(record_enumeration_offset_position - record_compaction_offset_position) AS total_uncompacted_depth FROM collection;
+SELECT CAST(COALESCE(SUM(record_enumeration_offset_position - record_compaction_offset_position), 0) AS bigint) AS total_uncompacted_depth FROM collection;
